+++ docs/user_management/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this gear are documented in this file.
 
 ## 1.4.9
 
+* Updates to pull REDCap API code from library instead
 * Fixes an error pulling the full list of coperson records when the record count
   is a multiple of the page size.